feat(Files): Renames files more properly.
