+++ src/data/models/character-schema.ts
