+++ src/features/commands/character/character-command.ts
