+++ src/types/character/stat-type.ts
