+++ src/types/shop/item-type.ts
@@ -1,4 +1,4 @@
-import { Effect } from '../skills/effects'
+import { Effect } from '../skills/effect-type'
 
 export type Item = {
   name: string,
