+++ src/types/skills/effect-type.ts
